Refactor renamed middleware.js and moved it to new utils folder and updated references to it
