+++ server/routes/cart-route.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { isLoggedIn } = require('../middleware');
+const { isLoggedIn } = require('../utils/auth-middleware');
 const { addProductToCart, fetchCartProducts} = require('../controllers/cart-controller')
 
 router.post('/:account_id',isLoggedIn,addProductToCart);
